Add an auto-fit points to box feature
Works as a combination of expand to box and shrink to points, with
a little bit of intelligence: It repeats expand a limited number
of times until there are no points nearby anymore. If that is not
the case, no auto fit is performed. This avoids growing into
neighboring objects or the ground.